+++ src/Compilers/CSharp/Portable/Symbols/Synthesized/GeneratedNameConstants.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+
+namespace Microsoft.CodeAnalysis.CSharp.Symbols
+{
+    internal static class GeneratedNameConstants
+    {
+        public const char DotReplacementInTypeNames = '-';
+        public const string SynthesizedLocalNamePrefix = "CS$";
+    }
+}
