+++ src/Compilers/CSharp/Portable/Symbols/Synthesized/GeneratedNames.cs
@@ -6,134 +6,12 @@
 
 using System;
 using System.Diagnostics;
-using System.Globalization;
-using System.Linq;
 using System.Text;
-using Microsoft.CodeAnalysis.Collections;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
 {
-    internal static class GeneratedNameConstants
-    {
-        public const char DotReplacementInTypeNames = '-';
-        public const string SynthesizedLocalNamePrefix = "CS$";
-    }
-
-    internal static class GeneratedNameParser
-    {
-        internal static bool IsSynthesizedLocalName(string name)
-            => name.StartsWith(GeneratedNameConstants.SynthesizedLocalNamePrefix, StringComparison.Ordinal);
-
-        // The type of generated name. See TryParseGeneratedName.
-        internal static GeneratedNameKind GetKind(string name)
-            => TryParseGeneratedName(name, out var kind, out _, out _) ? kind : GeneratedNameKind.None;
-
-        // Parse the generated name. Returns true for names of the form
-        // [CS$]<[middle]>c[__[suffix]] where [CS$] is included for certain
-        // generated names, where [middle] and [__[suffix]] are optional,
-        // and where c is a single character in [1-9a-z]
-        // (csharp\LanguageAnalysis\LIB\SpecialName.cpp).
-        internal static bool TryParseGeneratedName(
-            string name,
-            out GeneratedNameKind kind,
-            out int openBracketOffset,
-            out int closeBracketOffset)
-        {
-            openBracketOffset = -1;
-            if (name.StartsWith("CS$<", StringComparison.Ordinal))
-            {
-                openBracketOffset = 3;
-            }
-            else if (name.StartsWith("<", StringComparison.Ordinal))
-            {
-                openBracketOffset = 0;
-            }
-
-            if (openBracketOffset >= 0)
-            {
-                closeBracketOffset = name.IndexOfBalancedParenthesis(openBracketOffset, '>');
-                if (closeBracketOffset >= 0 && closeBracketOffset + 1 < name.Length)
-                {
-                    int c = name[closeBracketOffset + 1];
-                    if (c is >= '1' and <= '9' or >= 'a' and <= 'z') // Note '0' is not special.
-                    {
-                        kind = (GeneratedNameKind)c;
-                        return true;
-                    }
-                }
-            }
-
-            kind = GeneratedNameKind.None;
-            openBracketOffset = -1;
-            closeBracketOffset = -1;
-            return false;
-        }
-
-        internal static bool TryParseSourceMethodNameFromGeneratedName(string generatedName, GeneratedNameKind requiredKind, out string methodName)
-        {
-            int openBracketOffset;
-            int closeBracketOffset;
-            GeneratedNameKind kind;
-            if (!TryParseGeneratedName(generatedName, out kind, out openBracketOffset, out closeBracketOffset))
-            {
-                methodName = null;
-                return false;
-            }
-
-            if (requiredKind != 0 && kind != requiredKind)
-            {
-                methodName = null;
-                return false;
-            }
-
-            methodName = generatedName.Substring(openBracketOffset + 1, closeBracketOffset - openBracketOffset - 1);
-
-            if (kind.IsTypeName())
-            {
-                methodName = methodName.Replace(GeneratedNameConstants.DotReplacementInTypeNames, '.');
-            }
-
-            return true;
-        }
-
-        // Extracts the slot index from a name of a field that stores hoisted variables or awaiters.
-        // Such a name ends with "__{slot index + 1}". 
-        // Returned slot index is >= 0.
-        internal static bool TryParseSlotIndex(string fieldName, out int slotIndex)
-        {
-            int lastUnder = fieldName.LastIndexOf('_');
-            if (lastUnder - 1 < 0 || lastUnder == fieldName.Length || fieldName[lastUnder - 1] != '_')
-            {
-                slotIndex = -1;
-                return false;
-            }
-
-            if (int.TryParse(fieldName.Substring(lastUnder + 1), NumberStyles.None, CultureInfo.InvariantCulture, out slotIndex) && slotIndex >= 1)
-            {
-                slotIndex--;
-                return true;
-            }
-
-            slotIndex = -1;
-            return false;
-        }
-
-        internal static bool TryParseAnonymousTypeParameterName(string typeParameterName, out string propertyName)
-        {
-            if (typeParameterName.StartsWith("<", StringComparison.Ordinal) &&
-                typeParameterName.EndsWith(">j__TPar", StringComparison.Ordinal))
-            {
-                propertyName = typeParameterName.Substring(1, typeParameterName.Length - 9);
-                return true;
-            }
-
-            propertyName = null;
-            return false;
-        }
-    }
-
     internal static class GeneratedNames
     {
         private const string SuffixSeparator = "__";
